src/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
functions-framework==3.*
pytest==7.4.3
pulumi-gcp
google-cloud-secret-manager
